Add files via upload
The dataset is about the Share prices from October 2015 till April 2023 of two famous manufacturers of cool sports cars Ferrari and Tesla.
The Tableau visualization provides essential insights that can be useful for investors and analysts: –
1)	Total volume of Tesla shares showed exponential growth from 2015 to 2020 and again dropped drastically over the next 3 years. On the other hand, the total volume of Ferrari shares showed a decrease over the years and never went up.
2)	When Tesla’s grip on Market share was tightened its stock price increased sharply on the other hand, Ferrari failed to attract customers at the same pace resulting in the marginal decline of its stock price.
This information can be valuable for investors interested in buying and selling stocks of these companies. It provides an overview of how these two companies have performed in the stock market over the years and how their stock prices are affected by market share and customer attraction.